requirements.txt
fastapi==0.114.1
python-dotenv==1.0.1
uvicorn==0.30.6